moved buttons to template folder
