+++ client/src/components/ButtonTypes/GoToEntryButton/GoToEntryButton.js
@@ -3,7 +3,7 @@ import './GoToEntryButton.scss'
 import { useHistory, useParams } from "react-router-dom"
 
 
-import Button from '../../Button/Button';
+import Button from '../../../templates/Button/Button';
 
 const GoToEntryButton = props => {
