+++ client/src/components/ButtonTypes/LogOutButton/LogOutButton.js
@@ -3,7 +3,7 @@ import './LogOutButton.scss'
 import {connect} from 'react-redux'
 
 import {logOutAction} from '../../../redux/actions/logInOut'
-import Button from '../../Button/Button'
+import Button from '../../../templates/Button/Button'
 
 const LogOutButton = props => {
