+++ client/src/components/EntryCreateForm/SubmitCreate.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Button from '../Button/Button'
+import Button from '../../templates/Button/Button'
 
 import {connect} from 'react-redux'
