+++ client/src/controllers/ProfileEdit/ProfileEdit.js
@@ -4,7 +4,7 @@ import './ProfileEdit.scss'
 import {connect} from 'react-redux'
 import {editProfileAction} from '../../redux/actions/db_put'
 import {deleteProfileAction} from '../../redux/actions/profile'
-import Button from '../../components/Button/Button'
+import Button from '../../templates/Button/Button'
 
 const ProfileEdit = props => {
     const {user, editProfile, deleteProfile} = props
