+++ client/src/controllers/StoryCreate/StoryCreate.js
@@ -1,6 +1,6 @@
 import React, {useState} from 'react'
 import './StoryCreate.scss'
-import Button from '../../components/Button/Button'
+import Button from '../../templates/Button/Button'
 import {connect} from 'react-redux'
 import {createStoryAction} from '../../redux/actions/db_post'
