+++ client/src/controllers/StoryEdit/StoryEdit.js
@@ -5,7 +5,7 @@ import {connect} from 'react-redux'
 import {editStoryAction} from '../../redux/actions/db_put'
 import {deleteStoryAction} from '../../redux/actions/db_delete'
 
-import Button from '../../components/Button/Button'
+import Button from '../../templates/Button/Button'
 
 const StoryEdit = props => {
     const {story} = props
